Migrating to latest .NET version
