--- ShortherUrlCore/Business/IShortnerUrlBS.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace ShortherUrlCore.Business
-{
-    public interface IShortnerUrlBS
-    {
-        Task<string> Process(string originalUrl);
-    }
-}
