--- ShortherUrlCore/Models/ShortUrl.cs
@@ -1,8 +1,0 @@
-﻿namespace ShortherUrlCore.Models
-{
-    public class ShortUrl
-    {
-        public string OriginalUrl { get; set; }
-        public string ShortnedUrl { get; set; }
-    }
-}
