--- ShortherUrlCore/Storage/IStorageManager.cs
@@ -1,12 +1,0 @@
-﻿using ShortherUrlCore.Models;
-using System.Threading.Tasks;
-
-namespace ShortherUrlCore.Storage
-{
-    public interface IStorageManager
-    {
-        Task<ShortUrl> Get(string hashUrl);
-
-        Task Insert(ShortUrl shortUrl);
-    }
-}
