--- ShortherUrlCore/Storage/SqlServer/Models/ShortUrlSqlServer.cs
@@ -1,12 +1,0 @@
-﻿using Dapper.Contrib.Extensions;
-
-namespace ShortherUrlCore.Storage.SqlServer.Models
-{
-    [Table("ShortUrl")]
-    public class ShortUrlSqlServer
-    {
-        [ExplicitKey]
-        public string ShortnedUrl { get; set; }
-        public string OriginalUrl { get; set; }
-    }
-}
