--- ShortnerUrl/ShortnerUrl.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ShortherUrlCore\ShortherUrlCore.csproj" />
-  </ItemGroup>
-
-
-</Project>
