--- ShortnerUrl/Startup.cs
@@ -1,54 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using ShortherUrlCore.Business;
-using ShortherUrlCore.Storage;
-using ShortherUrlCore.Storage.AzureTableStorage;
-using ShortherUrlCore.Storage.InMemory;
-using ShortherUrlCore.Storage.SqlServer;
-
-namespace ShortnerUrl
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddControllers();
-
-            // Choose which Data Provider to use
-            services.AddSingleton<IStorageManager, InMemoryStorageManager>();
-            //services.AddSingleton<IStorageManager, AzureTableStorageManager>();
-            //services.AddSingleton<IStorageManager, SqlServerStorageManager>();
-
-            services.AddScoped<IShortnerUrlBS, ShortnerUrlBS>();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseRouting();
-
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
-        }
-    }
-}
