+++ UrlShortner.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35707.178 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UrlShortner", "UrlShortner\UrlShortner.csproj", "{FF524D92-71F5-4329-B195-893366A617B6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UrlShortnerCore", "UrlShortnerCore\UrlShortnerCore.csproj", "{EFF277A0-700D-47A5-847C-7D5645780366}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UrlShortnerTest", "UrlShortnerTest\UrlShortnerTest.csproj", "{10B75DB0-3688-49DA-95DF-BFC916BFBCE1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FF524D92-71F5-4329-B195-893366A617B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF524D92-71F5-4329-B195-893366A617B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FF524D92-71F5-4329-B195-893366A617B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF524D92-71F5-4329-B195-893366A617B6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EFF277A0-700D-47A5-847C-7D5645780366}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFF277A0-700D-47A5-847C-7D5645780366}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFF277A0-700D-47A5-847C-7D5645780366}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFF277A0-700D-47A5-847C-7D5645780366}.Release|Any CPU.Build.0 = Release|Any CPU
+		{10B75DB0-3688-49DA-95DF-BFC916BFBCE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10B75DB0-3688-49DA-95DF-BFC916BFBCE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10B75DB0-3688-49DA-95DF-BFC916BFBCE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10B75DB0-3688-49DA-95DF-BFC916BFBCE1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
