+++ UrlShortnerCore/Business/IShortnerUrlBS.cs
@@ -1,0 +1,7 @@
+﻿namespace ShortherUrlCore.Business
+{
+    public interface IShortnerUrlBS
+    {
+        Task<string> Process(string originalUrl);
+    }
+}
