+++ UrlShortnerCore/Storage/IStorageManager.cs
@@ -1,0 +1,11 @@
+﻿using ShortherUrlCore.Models;
+
+namespace ShortherUrlCore.Storage
+{
+    public interface IStorageManager
+    {
+        Task<ShortUrl?> Get(string hashUrl);
+
+        Task Insert(ShortUrl shortUrl);
+    }
+}
