+++ UrlShortnerCore/Storage/SqlServer/Models/ShortUrlSqlServer.cs
@@ -1,0 +1,12 @@
+﻿using Dapper.Contrib.Extensions;
+
+namespace ShortherUrlCore.Storage.SqlServer.Models
+{
+    [Table("ShortUrl")]
+    public class ShortUrlSqlServer
+    {
+        [ExplicitKey]
+        public required string ShortnedUrl { get; set; }
+        public required string OriginalUrl { get; set; }
+    }
+}
